Change `std::fmodf` to `std::modf` to fix gcc & clang compilation
Not exactly sure why, as `std::fmodf` should be available in c++20
(see https://en.cppreference.com/w/cpp/numeric/math/fmod), but it
appears to not exist...

+++ arkose/animation/Animation.cpp
@@ -105,7 +105,7 @@ Animation::SampledInputTrack Animation::evaluateInputTrack(size_t inputTrackIdx,
         while (inputTrackTime < inputTrackStart) {
             inputTrackTime += inputTrackLength;
         }
-        inputTrackTime = inputTrackStart + std::fmodf(inputTrackTime - inputTrackStart, inputTrackLength);
+        inputTrackTime = inputTrackStart + std::fmod(inputTrackTime - inputTrackStart, inputTrackLength);
     }
 
     // Find the first input track index
